+++ CHANGELOG.md
@@ -7,7 +7,8 @@ FEATURES:
 IMPROVEMENTS:
 
 * `vault_agent_registration`: Add support for `local` field. Requires Vault 2.2.0 or later. ([#3039](https://github.com/hashicorp/terraform-provider-vault/pull/3039))
-* `vault_oauth_resource_server_config_profile`: Add support for `local` field. Requires Vault 2.2.0 or later. ([#3039](https://github.com/hashicorp/terraform-provider-vault/pull/3039)) 
+* `vault_oauth_resource_server_config_profile`: Add support for `local` field. Requires Vault 2.2.0 or later. ([#3039](https://github.com/hashicorp/terraform-provider-vault/pull/3039))
+* `vault_plugin`: Add `reload` attribute to support reloading the plugin when it's updated. ([#3056](https://github.com/hashicorp/terraform-provider-vault/pull/3056))
 
 ## 5.12.0 (September 17, 2026)